More compiler warning fixes for GCC

+++ apps/parallel_hdf5/main.cpp
@@ -55,7 +55,9 @@ struct mpi_h5_reader_t
 
     explicit mpi_h5_reader_t(int size_, int rank_) : size{size_}, rank{rank_} {}
 
-    auto read_csr5(h5::group_view_t matrix_group) { return h5::load_csr5_partial(matrix_group, rank, size); }
+    auto read_csr5(h5::group_view_t matrix_group) {
+        return h5::load_csr5_partial(matrix_group, static_cast<size_t>(rank), static_cast<size_t>(size));
+    }
 };
 
 int main(int argc, char* argv[]) try {
